% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svPkgdown-package.R
\docType{package}
\name{svPkgdown-package}
\alias{svPkgdown}
\alias{svPkgdown-package}
\title{'pkgdown' Template for 'SciViews::R' Packages}
\description{
This package is used to produce the documentation of R packages on
https://www.sciviews.org. Please, do not use it unmodified for other purposes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SciViews/svPkgdown}
  \item \url{https://www.sciviews.org/svPkgdown/}
  \item \url{https://sciviews.r-universe.dev/svPkgdown}
  \item Report bugs at \url{https://github.com/SciViews/svPkgdown/issues}
}

}
\author{
\strong{Maintainer}: Philippe Grosjean \email{phgrosjean@sciviews.org} (\href{https://orcid.org/0000-0002-2694-9471}{ORCID})

}
