% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svPkgdown-package.R
\docType{package}
\name{svPkgdown-package}
\alias{svPkgdown}
\alias{svPkgdown-package}
\title{svPkgdown: 'pkgdown' Template for SciViews Packages}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

This is a private template to style pkgdown sub-sites in
  https://www.sciviews.org. Please don't use it for another purpose.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SciViews/svPkgdown}
  \item \url{https://www.sciviews.org/svPkgdown/}
  \item Report bugs at \url{https://github.com/SciViews/svPkgdown/issues}
}

}
\author{
\strong{Maintainer}: Philippe Grosjean \email{phgrosjean@sciviews.org} (\href{https://orcid.org/0000-0002-2694-9471}{ORCID})

}
\keyword{internal}
